added README; added datasheets; changed config.yaml parameters, mainly bmp280 mode changed from forced to normal and output data rate directly controlled by oversampling setting

+++ include/sen0140_ros2/adxl345.hpp
@@ -33,7 +33,7 @@ public:
     ~Adxl345();
 
     void initialize(
-        int odr_hz,
+        double odr_hz,
         AccelRange range);
 
     bool is_data_ready();
@@ -42,7 +42,7 @@ public:
 
 private:
     void configure(
-        int odr_hz,
+        double odr_hz,
         AccelRange range);
 
     void write_register(
